Microsoft June 2021 Game Development Kit (GDK) Public RELEASE Build [210600 ]
This is the official General Availability (GA) and first public release of the Microsoft Game Development Kit (GDK).

Build version nomenclature will show up on binaries like this:  210600
21 = Year (2021)
06 = Month (June)

There are 3 major GDK releases a year (Spring, Summer, and Fall) which can contain both features and fixes.

There are GDK updates in between  which are focused on stability and fixes.