June 2021 GDK Readme Updates
This section is under development as we take feedback from the community what information is most useful to present up-front vs. in the documentation (https://aka.ms/gamedevdocs).

There are two distinct audiences for the GDK (developers already in or onboarding into the Xbox ecosystem and developers not yet / interested in the Xbox ecosystem).  As the distinct needs of these audiences are more clearly identified, the Readme.MD will serve to help provide clarity and direction on where to get started depending on where a developer is at in their journey to build games for the Xbox App on Windows, Xbox Game Pass for PC, Xbox consoles (Xbox Series X|S, Xbox One), or Xbox Cloud Gaming (game streaming).