Public GDK Docs Test
The public GDK docs are published to docs.microsoft.com (DMC) on https://aka.ms/gamedevdocs.

However, we are looking at developer requests to consolidate on GitHub to reduce the number of places developers have to look to find development resources.  Please refer to the link above for latest and greatest until the Readme is updated to reflect that GitHub pages are our default publishing mode.