Add News Section to Readme.md
The GDK uses a multi-headed branch model where developers can choose either latest and greatest MAJOR releases or stay on a branch and cherry pick which UPDATE releases meet their needs.  However, this means that UPDATE releases can appear chronologically newer than the latest and greates MAJOR release.  At this time we are not manipulating GIT_COMMITER_DATE to preserve the order in which kits are published, so this relnote is needed.